Fix for BackFromPreviewLosesText and RefreshEditPage.  Based on code running for a couple of months locally, and tested on IE5.5, Opera 5.12 and Mozilla 0.9.9.  Adds a new routine, for use by 'edit' only at present, that sets additional cache-control headers that cause IE5 and IE6 to cache the Edit page when hitting Back button in browser.
NOTE: This currently disables the call to Plugins::writeHeaderHandler -
this is a temporary measure.  Even though we could put all
the code in a plugin, the bug fix really needs to be in the core code,
because IE5 is so common.  Also, any plugin that does cache control
for the Edit page should have these headers overridden to avoid re-creating
this bug. See BackFromPreviewLosesText for more discussion.


git-svn-id: http://develop.twiki.org/svn/trunk@791 a00a5322-12db-0310-a70b-8735589c885e