Initial revision of JavaPaste.
JavaPaste is a drag and drop java applet that, when a set of files is dragged onto it, calls upload.pl.
It also uploads files such as pictures when these are pasted into the Applet.

Work originally authored by Catherine Macleod.
Drag and drop functionality and bug fixes coded by Main.RichardEllery
Initial idea and lets-get-it-out-the-door motivation by Main.MartinCleaver,

This code is still alpha quality. Caveat Emptor.


git-svn-id: http://develop.twiki.org/svn/twiki/branches/DEVELOP@4691 a00a5322-12db-0310-a70b-8735589c885e