RegisterCgiScriptRewrite: MERGE TIME
It turns out that neither Crawford nor I are able to work on the code over the next week yet we have both had 
large changesets outstanding. As my stuff is largely isolated to Register.pm and RegisterTestSet.pm, and is
not too disruptive when broken we've agreed that it is better to get it in where people can look and help 
troubleshoot.

There is something broken on my build - but as the tests worked fine before I merged, I feel it must be a
config setting somewhere as even view fails. All the compile tests succeed.

I apologise for leaving it in this state - I hope to have time to look on Saturday afternoon.

PS. I didn't really change all those files in the TWiki web - this is a consequence of the lineending problem.

M.



git-svn-id: http://develop.twiki.org/svn/twiki/branches/DEVELOP@3393 a00a5322-12db-0310-a70b-8735589c885e