Item4621: don't store auto-created images in pub
Storing thumbnails in pub may crash if some of the webs are shared on a
read-only filesystem among several instances of TWiki.  The workarea is no
option anymore, as it is outside of the pub tree and thus not visible to
apache. So ImageGalleryPlugin will store it in pub/images/... now.



git-svn-id: http://develop.twiki.org/svn/twiki/branches/MAIN/twikiplugins@14855 a00a5322-12db-0310-a70b-8735589c885e