Item5157:
Added LdapApacheLogin manager to be used in a setting where you authenticate
using apache and you would have been using the standard ApacheLogin manager.
LdapApacheLogin manager is a drop in replacement for it, keeping track of new
users and granting them to be known to the cache in addition. 

More fixes: 
   * made cache updates quasi atomic so a connection failure to the LDAP
     directory won't empty the cache 
   * refined internal api to update a single cache entry 
   * added docu using LdapContrib in an SSO setup



git-svn-id: http://develop.twiki.org/svn/twiki/branches/MAIN/twikiplugins@16075 a00a5322-12db-0310-a70b-8735589c885e