Item6088: The Perl modules do not have fully qualified names and wrong convention/or error in the name of the file/module name.
Reverting changes for Wysiwyg Plugin code. The version r17714 is reverted to r17701.

The Item6088:  The Perl modules do not have fully qualified names and wrong convention/or error in the name of the file/module name - corrected this and verified on a few instances of TWiki

is Open again. This will be fixed by considering changes of r17701 version.

--Sopan Shewale



git-svn-id: http://develop.twiki.org/svn/twiki/trunk@17720 a00a5322-12db-0310-a70b-8735589c885e